backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain
langchain-openai
python-dotenv
colorama
